:sparkles: feat(frontend): 공통 컴포넌트 schedule-container 추가 (#72)
