+++ frontend/components/screens/feed/feed-container/feed-container.interface.ts
@@ -1,0 +1,4 @@
+export interface FeedContainerProps {
+	options: 'TOP' | 'MYFEED' | 'ALL';
+	handleIsLottie: (status: boolean) => void;
+}
