+++ frontend/components/screens/schedule/schedule-container/ScheduleContainer.module.scss
@@ -1,0 +1,6 @@
+.schedule_container {
+	@screen md {
+		@apply mx-0;
+	}
+	@apply pt-6 pb-8 flex flex-col gap-10 mx-4;
+}
