+++ frontend/components/screens/schedule/schedule-container/schedule-container.interface.ts
@@ -1,0 +1,3 @@
+export interface ScheduleContainerProps {
+	options: 'SCHEDULEALL' | 'MYSCHEDULE' | 'SHAREDSCHEDULE';
+}
